+++ src/index.ts
@@ -1,2 +1,3 @@
 export { Store, Subscription } from "./Store"
 export { System } from "./System"
+export { ActorReceiveBuilder as ReceiveBuilder } from "js-actor"